requirements.txt
Django
spacy
PyPDF
openai
textract
python-dotenv
llama-index
PyMuPDF
llama-index-vector-stores-chroma
llama-index-llms-openai
llama-index-embeddings-openai